Updated Navigation and Network DI bind and provide modules.
